move BoxedStateStore to a new file
