+++ etl/src/v2/state/store/base.rs
@@ -1,5 +1,5 @@
 use postgres::schema::TableId;
-use std::{collections::HashMap, sync::Arc};
+use std::collections::HashMap;
 use thiserror::Error;
 
 use crate::v2::{
@@ -45,49 +45,4 @@ pub trait StateStore: Send + Sync {
         table_id: TableId,
         state: TableReplicationPhase,
     ) -> Result<(), StateStoreError>;
-}
-
-#[derive(Clone)]
-pub struct BoxedStateStore {
-    inner: Arc<dyn StateStore + Send + Sync>,
-}
-
-impl BoxedStateStore {
-    pub fn new<S: StateStore + Send + Sync + 'static>(store: S) -> Self {
-        BoxedStateStore {
-            inner: Arc::new(store),
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl StateStore for BoxedStateStore {
-    async fn get_table_replication_state(
-        &self,
-        table_id: TableId,
-    ) -> Result<Option<TableReplicationPhase>, StateStoreError> {
-        self.inner.get_table_replication_state(table_id).await
-    }
-
-    async fn get_table_replication_states(
-        &self,
-    ) -> Result<HashMap<TableId, TableReplicationPhase>, StateStoreError> {
-        self.inner.get_table_replication_states().await
-    }
-
-    async fn load_table_replication_states(
-        &self,
-    ) -> Result<HashMap<TableId, TableReplicationPhase>, StateStoreError> {
-        self.inner.load_table_replication_states().await
-    }
-
-    async fn store_table_replication_state(
-        &self,
-        table_id: TableId,
-        state: TableReplicationPhase,
-    ) -> Result<(), StateStoreError> {
-        self.inner
-            .store_table_replication_state(table_id, state)
-            .await
-    }
 }
