+++ etl/src/v2/state/store/mod.rs
@@ -1,3 +1,4 @@
 pub mod base;
+pub mod boxed;
 pub mod memory;
 pub mod postgres;
