+++ replicator/src/core.rs
@@ -5,7 +5,8 @@ use etl::v2::config::retry::RetryConfig;
 use etl::v2::destination::base::Destination;
 use etl::v2::destination::memory::MemoryDestination;
 use etl::v2::pipeline::{Pipeline, PipelineIdentity};
-use etl::v2::state::store::base::{BoxedStateStore, StateStore};
+use etl::v2::state::store::base::StateStore;
+use etl::v2::state::store::boxed::BoxedStateStore;
 use etl::v2::state::store::memory::MemoryStateStore;
 use etl::v2::state::store::postgres::PostgresStateStore;
 use etl::SslMode;
